Corrige quelques leaks

+++ inc/config.h
@@ -2,8 +2,8 @@
 #define CONFIG__H
 
 struct config {
-    int DAEMON_WORKER_MAX;
-    int REQUEST_QUEUE_MAX;
+    size_t DAEMON_WORKER_MAX;
+    size_t REQUEST_QUEUE_MAX;
 };
 
 int config_load(struct config *ptr, const char *filename);
